dev-requirements.txt
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.3.2
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
Markdown==3.0.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
toml==0.10.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
